pyproject.toml
[tool.poetry]
name = "project_name"
version = "0.0.1"
description = ""
authors = ["mishaga"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
fastapi = "^0.110.2"
uvicorn = "^0.29.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.7.0"}
pydantic-settings = "^2.2.1"
pyjwt = "^2.8.0"
sentry-sdk = "^1.45.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-env = "^1.1.1"
pytest-asyncio = "^0.23.6"
faker = "^24.11.0"
wemake-python-styleguide = "^0.19.2"
flake8-pyproject = "^1.2.3"
alembic-autogen-check = "^1.1.1"
dirty-equals = "^0.7.1"
httpx = "^0.27.0"
jsonschema = "^4.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:cacheprovider"


[tool.pytest_env]
ENVIRONMENT = "tests"
UNIT_TESTS = "true"
POSTGRES_DATABASE = "database_tests"


[tool.isort]
profile = "wemake"
known_first_party = ["api", "common_functions", "config", "db", "tests"]
skip_glob = ["db/migrations/*"]


[tool.flake8]
format = "wemake"
max-complexity = "10"
import-order-style = "pycharm"
min-coverage-percents = "80"
max-line-length = "120"
application-import-names = ["api", "common_functions", "config", "db", "tests"]
class_attributes_order = [
    "field",
    "meta_class",
    "nested_class",
    "magic_method",
    "property_method",
    "static_method",
    "class_method",
    "method",
    "private_method",
]
exclude = ["db/migrations/*.py", "management_commands/*.py"]
ignore = ["D100", "D104", "DAR101", "DAR201", "DAR301", "DAR401", "WPS305", "WPS306", "WPS323"]
per-file-ignores = [
    "*/__init__.py: F401, WPS412",
    "api/*.py: B008, WPS404",
    "config/const.py: WPS432",
    "tests/*.py: WPS226, WPS316, WPS520, S101",
    "tests/fixtures/*.py: WPS442",
]
# B001   – Do not use bare except
# B008   – Do not perform function calls in argument defaults
# D100   – Missing docstring in public module
# D104   – Missing docstring in public package
# DAR101 – Missing parameter(s) in Docstring
# DAR201 – Missing "Returns" in Docstring
# DAR301 – Missing "Yields" in Docstring
# DAR401 – Missing exception in Raises section
# E722   – Do not use bare except
# F401   – Imported but unused
# E711   – comparison to None
# N805   – first argument of a method should be named 'self'
# S101   – Use of assert detected
# WPS110 – Found wrong variable name
# WPS115 – Found upper-case constant in a class
# WPS201 – Found module with too many imports
# WPS202 – Found too many module members
# WPS204 – Found overused expression
# WPS210 – Found too many local variables
# WPS213 – Found too many expressions
# WPS214 – Found too many methods
# WPS217 – Found too many await expressions
# WPS218 – Found too many `assert` statements
# WPS221 – Found line with high Jones Complexity
# WPS226 – Found string literal over-use
# WPS232 – Found module cognitive complexity that is too high
# WPS300 – Found local folder import
# WPS301 – Found dotted raw import
# WPS305 – Found f-string
# WPS306 – Found class without a base class
# WPS316 – Found context manager with too many assignments
# WPS318 – Found extra indentation
# WPS323 – Found % string formatting
# WPS404 – Found complex default value
# WPS412 – Found `__init__.py` module with logic
# WPS432 – Found magic number
# WPS442 – Found outer scope names shadowing
# WPS450 – Found protected object import
# WPS451 – Found positional-only argument
# WPS520 – Found compare with falsy constant
# Q001   – Single quote multiline found but double quotes preferred
